provisioning_service/requirements.txt
Flask-Cors==4.0.0
bcrypt==4.1.2
blinker==1.9.0
boto3==1.34.34
botocore==1.34.162
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
Flask==3.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
google-auth==2.49.1
greenlet==3.3.2
gunicorn==21.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
kubernetes==28.1.0
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
psycopg2-binary==2.9.9
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
PyYAML==6.0.3
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==1.26.20
websocket-client==1.9.0
Werkzeug==3.1.6
flask-sock==0.7.0
simple-websocket==1.1.0
gevent==24.2.1
